+++ gwcelery/tests/test_tasks_superevents.py
@@ -41,6 +41,12 @@ class SingleT0212TTPResponse(object):
         response = resource_json(
             __name__, 'data/T0212_S0039_preferred.json')
         response['instruments'] = "H1"
+        # FIXME setting the chisq value for None for other detetors
+        # this is supposed to be temporary fix. To be removed when
+        # gstlal uploads correct 'instruments' field
+        response['extra_attributes']['SingleInspiral'] = [
+            {'chisq': 1 if instrument in ['H1'] else None}
+            for instrument in ['H1', 'L1', 'V1']]
         return response
 
 
@@ -87,7 +93,9 @@ def test_update_preferred_event(mock_db):
                     superevent="some_superevent",
                     far=1e-30,
                     snr=30.0)
-    with patch.object(gracedb.client, 'updateSuperevent') as p:
+    with patch.object(gracedb.client, 'updateSuperevent') as p, \
+            patch('gwcelery.tasks.gracedb.get_number_of_instruments',
+                  return_value=5):
         superevents._update_superevent('S0039',
                                        'T0212',
                                        mock_response(),
@@ -154,7 +162,9 @@ def test_upload_same_event():
         }
     }
     with patch.object(gracedb.client, 'addEventToSuperevent') as p1, \
-            patch.object(gracedb.client, 'updateSuperevent') as p2:
+            patch.object(gracedb.client, 'updateSuperevent') as p2, \
+            patch('gwcelery.tasks.gracedb.get_number_of_instruments',
+                  return_value=2):
         superevents.handle(payload)
         p1.assert_called_once()
         p2.assert_not_called()
@@ -166,6 +176,7 @@ def test_parse_trigger_raising():
         superevents.handle(garbage_payload)
 
 
+@patch('gwcelery.tasks.gracedb.get_number_of_instruments', return_value=2)
 def test_parse_trigger_cbc_1(mock_db):
     """New trigger G000000, less significant than already
     existing superevent. Superevent window much larger than event
@@ -191,6 +202,7 @@ def test_parse_trigger_cbc_1(mock_db):
         p2.assert_not_called()
 
 
+@patch('gwcelery.tasks.gracedb.get_number_of_instruments', return_value=2)
 def test_parse_trigger_cbc_2(mock_db):
     """New trigger G000003, more significant than already
     existing superevent. Superevent window is much larger that
@@ -259,6 +271,7 @@ def test_parse_trigger_cbc_4(mock_db):
         p2.assert_not_called()
 
 
+@patch('gwcelery.tasks.gracedb.get_number_of_instruments', return_value=2)
 def test_parse_trigger_burst_1(mock_db):
     """New cwb trigger G000005 with gpstime lying partially in
     S0039 window, not more significant than existing preferred
@@ -289,6 +302,7 @@ def test_parse_trigger_burst_1(mock_db):
                                    t_start=pytest.approx(1163905214, abs=1))
 
 
+@patch('gwcelery.tasks.gracedb.get_number_of_instruments', return_value=2)
 def test_parse_trigger_burst_2(mock_db):
     """New oLIB trigger G000006 with gpstime lying partially in
     S0039 window, more significant than already existing preferred
@@ -373,6 +387,7 @@ def test_parse_trigger_burst_4(mock_db):
         p.assert_called_once()
 
 
+@patch('gwcelery.tasks.gracedb.get_number_of_instruments', return_value=1)
 def test_single_ifo_1(mock_db):
     """New single IFO trigger G000009 same event attributes as
     G000000 (except single IFO). Will be added to superevent"""
@@ -398,6 +413,8 @@ def test_single_ifo_1(mock_db):
         p2.assert_not_called()
 
 
+@patch('gwcelery.tasks.gracedb.get_number_of_instruments',
+       lambda _: 1 if _ == 'G000010' else 2)
 def test_single_ifo_2(mock_db):
     """New single IFO trigger G000010 same event attributed as G000003
     (except single IFO). More significant than existing superevent, but
@@ -425,6 +442,8 @@ def test_single_ifo_2(mock_db):
         p2.assert_not_called()
 
 
+@patch('gwcelery.tasks.gracedb.get_number_of_instruments',
+       lambda _: 2 if _ == 'G000011' else 1)
 def test_single_ifo_3(mock_db):
     """New multi-IFO trigger CBC trigger G000011. Existing
     preferred event is a single IFO. Preferred event
@@ -457,6 +476,8 @@ def test_single_ifo_3(mock_db):
             t_start=pytest.approx(1163905213.44, abs=1e-3))
 
 
+@patch('gwcelery.tasks.gracedb.get_number_of_instruments',
+       lambda _: 1 if _ == 'G000013' else 2)
 def test_single_ifo_4(mock_db):
     """Preferred event in superevent is a Multi-IFO Burst,
     new trigger is a CBC single. Preferred event is not updated